consolidate specs
